Add files via upload
DeleteString is a function that simply asks how many characters are in the character array entered by the user. Then it asks how long that array is (including all characters plus the NULL, I believe.) Finally, It asks for a start index. The start index is very interesting, as everything after the start index is deleted by the DeleteString function, as is apparent upon a cout of the modified user_word variable.