Practice with Pointers_1_2 - 12-28-2019.cc
This file builds a function called StrCpy. StryCpy, like the traditional strcpy, copies characters from source to destination. 

The main feature of the StrCpy function is this:

"while(*dest++ = *src++);"

Notice the semicolon at the end of the while loop. Because the assignment operator is used within the condition, the control variable is hanging within the loop. 

Because the null value equates to zero, and zero represents false, as soon as null is copied from src to dest, the loop will terminate.