Practice with Pointers_2_2 - Jan 2 2020.cc
This program passes a character type pointer to a function by reference. A buffer is used to create a temporary local array to receive user input; The "new" keyword is then used to dynamically create an array exactly big enough for the string that was entered, plus null. The string is then copied from buffer to the new string.