Tokenize user input.
This program tokenizes a user-input string that is populated after being passed-by-pointer to a nested function. That is, a nested function is passing a pointer by pointer. This program also uses dynamic memory allocation for handling the user-input array. It would be neat to dynamically allocate the p_token array's memory in a future iteration. In this version the universal variable BUFFER_LENGTH is used for defining the char buffer array p_token within GetString...